11.24 임동혁 다시 백업

+++ easyway/src/main/java/org/easyway/domain/office/AnnualVacation.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AnnualVacation {
 	
-	private int av_id;
+	private Long av_id;
 	private int av_annual;//몇년차인지
 	private int office_id;	
 }
